requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Transformers and model weights
transformers>=4.36.0
peft>=0.7.0              # LoRA (used in SFT regime)

# Vision-language alignment
open-clip-torch>=2.20.0

# Retrieval
faiss-gpu>=1.7.4         # use faiss-cpu if no GPU available

# Data / numerical
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# NLP utilities
nltk>=3.8.0

# Configuration management
omegaconf>=2.3.0

# Progress bars
tqdm>=4.65.0

# Distributed training
# (torch.distributed is bundled with PyTorch)

# Evaluation — clinical metrics
# CheXBert: install separately from https://github.com/stanfordmlgroup/CheXBert
# RadGraph: install separately from https://github.com/jackboyla/radgraph
